Adjusted the .bat file to target the correct folder in the D: drive (my bad for copy pasting earlier).
And made tick icon next to "check farms" in farm widget.


Next step would be to find a way to make the tick icon have a cross icon whenever the weather and farm data does not have an equal number of rows.
I noticed a function that was written in one of the files that "stretches" the rows of partial datasets but will need to look more into that before coming to a proper understanding of how this works.